Use gross amount in order for return items
I believe the notion of `pre_tax_amount` is only in the return items because
someone from a sales tax country wanted to use something with additional taxes.

This leads to unfortunate behaviour in VAT stores: While every other price or amount
is entered including VAT, for returns a store admin has to enter the return amount
excluding VAT. For most normal people, that's a complex, error-prone op, and one we
shouldn't make them do.

Unfortunately, this entails a lot of naming changes.