Refactor line_item.rb
This just moves lines up and down so it's easier to see where
associations, AR callbacks, configuration and actual methods are.

I've also moved to methods to private which are not called from
outside the line item instance. I think they were mostly public because
their tests required that, so I reorganized the tests for them a little.

I've also de-indented the private methods. Hound will be happy!

I'd like to have this in as preparation for the VAT work.