Removes devise initializer from dummy generator
The dummy app generator adds a devise initializer that's disabled in most of the time.

The official solidus_auth_devise gem has it's own devise initializer, that never gets generated in the dummy app, because the generator skips existing devise initializers (what's good, to prevent overwriting in the real host app).