Add spec for a foreign address with no zone
If you ship to a foreign country with no zone set, at the moment
you get the tax rates for the default zone (so if you're shipping to
Vanuatu, and there's no zone for that, you might get German taxes for
that order).

This commit introduces specs for this case (and sets them all to pending).