Remove redundant spec that contains an invalid initial order state
This spec had an initial order state ("complete") that was the same as the outcome order state. The test passed, but for the wrong reason. After setting the initial state to "confirm", the spec failed as the API no longer allows transitioning order state to complete using a PUT on the order. You must instead call the "complete" action, for which there is already a spec.