Change requirements for rejecting an option value
Before if either field was blank it would be rejected by the option type
model.  This led to a confusing behaviour where you could have one field
filled in and the other blank and try to update and be told it updated
successfully while it lost the option type you had started without
telling you that.  Now this will throw an error if you try that telling
you that the option type did not save.  However if both fields are blank
it will just reject the option type.