Add Configuration for Admin VAT country
This configuration setting, which is valid for the entire Solidus
store instance, will tell Solidus which VATs are included in prices
entered in the backend.

I am splitting up the meaning of Spree::Zone.default_tax here, because
that setting conflates two questions:
1. Are we entering backend prices including VAT?
2. Should carts without a tax address be taxed?

The second setting should be `Spree::Store`-specific.