Use Spree::Store.default_cart_tax_location in Spree::Order#tax_address
Yet another step in removing the default tax zone: When the order does not
have a tax_address, refer to the store's default cart tax location.

This required quite a few spec setup tweaks as it's the first time
I'm actually using the mandatory store on an order, , but nothing grave.