Add settings tab item helper function
This is very similar to the configuration menu item helper function but
that function name is specific to the configuration menu which we are
moving away from and lacks a datahook.  Also this will add the 'fa'
class to main consistency with the view of the tabs in the order view
page.