Use conditional relationship rather than manual accessor.
This is basically just a change in implementation and no change in
functionality as it still does the same thing. `Spree::User#default_address`
still returns the related default address.

Why I prefer this implementation is it allows the address to be
preloaded preventing O(N) queries if we need to print some address
info out about each user (i.e. name, phone, etc). Printing the
name for each user seems like a common operation so it seems it
would be good to provide a way to do that for a list of users
efficiently.

The `default_address=` implementation is remaining and overriding the
implmenetation the relationship would provide. This will allow the
special handling of address assignment to remain.