Fix fighting between skeleton and bootstrap grids
Previously there were some minor layout issues due to both grids
claiming ownership of the "container" and "row" classes. Skeleton sets
container to a fixed width 960px, where bootstrap by default uses a
max-width of 940px. Bootstrap also provides styles for "row", as part of
its grid system.

This disables the generation of grid classes in bootstrap to avoid the
mixed styling of "row". To make use of the bootstrap grids we will be
using mixins.

This changes the max-size of the container in bootstrap to 960px, so
that containers use the existing size defined by skeleton.