Replace ul/li-s with a div, place weight last
This commit replaces use of ul/li-s on the product form with plain
divs, achieving a much nicer visual layout for the product height/
width/depth/weight fields, which were previously clunky and misaligned.

The weight field has been moved from the 1st to the 4th position,
for improved intuitiveness in the new layout.