Make the reference to Spree::Store explicit in app config.
This appeared to be the source of a code reloading issue when this logic
was called. I've patched it by simply removing the deprecated logic, but
adding the spree namespace here seemed to improve the rails constant
autoloading issues around this particular part of the code.

The error I saw was: ArgumentError: A copy of Spree::AppConfiguration
was removed from the module tree but is still active. We narrowed this
down to the fact rails may have been looking for the constant on
AppConfiguration and using a more qualified name would make this more
robust.