Combine Order#token_based? and Gateway#supports?
`Order#token_based?` was a private method and was only used inside
`Order#handle_payment_preconditions` and only in conjunction with
the `supports?` method.  And the `supports?` method itself is only
used in that same spot.  So as best we could tell, `token_based?` was
being used to modify how `supports?` works.  Combining them seems more
straightforward.

The motivation here is to make it easier to support non-credit-card
payment sources reducing complexity and reducing the number of methods
that a non-credit-card source is required to implement.