Deprecate the price modifier system
This commit tells users not to use the price modifier system. The Deprecation is run from
Spree::Variant, not from the `PriceModifier` pricer. The method will not be called (and not
issue a Deprecation warning) if

- There are no options when creating the line item
- The options hash contains a key `:price` which is a `Numeric`

In the latter case, `Spree::LineItem will set the price independent of any pricer and thus does not need to call it.