Variant#default_price delegates: Split setters and getters
This is preliminary work to ease the transition to superseding prices.
Setting the price on a variant should in the future create a new price,
while getting the price should get it from an existing one.

This commit just splits the `delegate` statement in two, one for
setters and one for getters.