Deprecate Spree::Money.new with parsed string
Previously Spree::Money could be initialized with a string which would
be parsed, like "$1.00 USD" or "$2,10 EUR". This introduces a
Spree::Money.parse which should be used instead for this type of input.

This is done to more closely match the ::Money API, as well as to be
faster and less error-prone.

This wasn't done primarily for perforamnce, but here are benchmarks for
comparison:

           Money.new    592.309k (± 3.3%) i/s -      2.975M in 5.028945s
Monetize.from_string    148.176k (± 3.4%) i/s -    742.504k in 5.016709s
Monetize.from_numeric
                        115.933k (± 6.5%) i/s -    578.655k in 5.017139s
Monetize.from_bigdecimal
                        183.085k (± 2.8%) i/s -    919.242k in 5.024655s
      Monetize.parse     44.864k (± 4.5%) i/s -    227.988k in 5.092225s

Conflicts:
	core/lib/spree/money.rb