product_picker.js sku lookup actually works
This is used in the backend UI for picking products for
Spree::Promotion::Rules::Product, might be used other places too.

There was code in there that looked like someone was trying to
let the query match a product name or sku. But it totally didn't work,
didn't do anything, was wrong in two different ways.

Took me a couple hours to figure out what was going on, so I figured
I'd share the fix. Sorry, don't know how to test this, looks like
there aren't any tests for this kind of JS yet?