Always show the page header
There will now *always* be something we can show as the page title, so
the header will no long be "an optional thing that just happens to
always be there".

The header is a big part of our layout, it should be clear that it is
always going to be present.