Raise on deprecations in the test app
We want to know when we use things that we deprecate. This commit changes ActiveSupport::Deprecation
as well as Spree::Deprecation so that they raise an error instead of printing to stderr when
they come across a deprecated things in the test suite.

It also fixes a non-silenced deprecation warning in the country spec, as well as the
reimbursement factory which gets initialized with a total of `nil`, raising a deprecation
warning from Spree::Money when rendering the reimbursement email.