Always validate count_on_hand of StockItem
Previously, count_on_hand was only validated if it was changed and
specifically if it was changed to a lower number than it was previously.

This isn't a good approach. A record should be valid based on it's
current attributes. It shouldn't be considered valid just because it's
"more" or "less" valid than before. The existing code would allow, for
example, creating a valid backorderable StockItem with a negative
count_on_hand, and then setting it to be backorderable=false. This is an
invalid state but would be allowed because count_on_hand wasn't changed.

This commit just changes the validation to just always run when
backorderable=false.