Hide clear selection if no clear option.
select2 has an option called allowClear which allows the user to clear
the selection if there is a option with a value of "". Due to a conflict
between select2 and fontawesome the clear button was always showing even
if there was no blank option.

The cascade of styles is:

1. select2 hides the `abbr` tag by default[1]
2. If the `allowclear` class is added it is re-added [2]
3. Solidus' select2 overrides mixed in font-awesome styles [3]
4. This mixed in style made it visible again regardless of allowclear [4]

This commit fixes the problem by re-hiding the element after mixing in
font-awesome. This makes the default hidden and if allowclear is set
(only true when a blank option exists) then it will redisplay.

[1] https://github.com/select2/select2/blob/3.5.1/select2.css#L97
[2] https://github.com/select2/select2/blob/3.5.1/select2.css#L114
[3] https://github.com/solidusio/solidus/blob/master/backend/app/assets/stylesheets/spree/backend/plugins/_select2.scss#L31
[4] https://github.com/FortAwesome/Font-Awesome/blob/master/css/font-awesome.css#L15