Add the ability to simple results from taxon API
The taxon products API which is used to manage display order falls down
under its own weight in any sort of reasonable data set. In order to
simplify that, I have added the ability to customize the return data
from this controller.

In a simple case, the results will be the same as they are right now,
however you can pass `simple: 1` to the taxon/products API request and
it will return significantly less data.

Additionally, this updates the backend taxons.js.coffee to use this
simple view for all it needs. This reduces the amount of data that needs
to be queried and returned in order to make this fast. Previously I was
seeing 20+ second response times for a small number of products in a
complicated taxon structure. This reduces that down to the 100ms range.