Add paginate and default_per_page extension points
These utility methods can be used in API endpoints that use Kaminari for
pagination, providing a useful extension point for customized pagination
patterns.

For example, in a store that has products with many taxons, Kaminari's
`default_per_page` setting of 25 can be restricting, and prevent all
taxons from being shown in the solidus_backend product admin views. By
providing a `default_per_page` private method as an extension point,
this default can be overwritten specifically for the TaxonsController.

A further commit will update API controllers as needed to take advantage
of this strategy.