Merge tax updating code and deprecate Order#create_tax_charge!
Previously we had two different sets of tax updating code.  One was
Order#create_tax_charge! and the other was the code in OrderUpdater.
The effect of the former encompasses the effect of the latter, so
functionally we can merge these.

This may result in increased load since we're now using the more
encompassing method which was more heavyweight.