Update CC brand detection regex
The existing hash was incomplete, missing some brands. It
was also missing MasterCard's new (October 2016) BIN range
(222100-272099).

This hash was originally added in
https://github.com/spree/spree/pull/3918
and has not been changed since. I don't know the original source, but I
would rather trust the Hash taken from ActiveMerchant.

One notable difference is that this will treat maestro and mastercard as
separate brands.

NB: For most stores, this code will not be run, and instead
jquery.payment on the frontend will be responsible for providing the
credit card brand.