Improve readability of try_type_from_number
The existing code returned from within a find, and if the find ever
succeeded, would return '' because nil (the return of the find) .to_s is
''.

This works, but is difficult to read. I prefer being explicit about the
return value being an empty string if there isn't an early return.