Make PaymentsHelper available to Order views
The `confirm` screen shows the payments info[1]. This uses[2] the helper `payment_method_name`[3] without
explicitly indicating that controller is using that helper.

In many Rails apps, this may be ok due to `config.action_controller.include_all_helpers` defaulting to `true`. But
if the application has configured this to be `false` the confirm screen breaks. I am personally a fan of the old Rails
behavior (i.e. don't put all helpers in a big pile but make separate helper modules actually mean something). Even
though it is the old behavior it is still supported behavior by Rails so it seems an effort to work either way should
be made.

1. https://github.com/solidusio/solidus/blob/master/backend/app/views/spree/admin/orders/confirm.html.erb#L54
2. https://github.com/solidusio/solidus/blob/master/backend/app/views/spree/admin/payments/_list.html.erb#L22
3. https://github.com/solidusio/solidus/blob/master/backend/app/helpers/spree/admin/payments_helper.rb#L4