Remove unused nested-attribute.js
This javascript was intended to be used by the generate_template and
remove_nested helpers which were removed in the previous commit.

Any code which used this has been gone since 2013
cc1d72874b75ced3dacfa56d944ca48dfe4b0968.

It looked like this was being used by some of the promotion rules, but
in fact wasn't because the code was only bound to elements which existed
on DOM ready (promotion rules are rendered by JS later).

Tested by running the following JS on all page loads

    if($('.remove').length) {
      console.log('Found on ' + document.location.href)
    }