Refactor order declaration in qty-adjustment spec
This change restructure the usage and declaration of `order` and
`line_items` in the related spec. Current setup with the basic order
factory result in an order in an incorrect state because `item_total`
and `item_count` are not updated.

Since the `tiered_percent` calculator depends on the order_total, we need
to ensure the order is properly updated before using it, and relying on
`order.reload.updater.update` is brittle and error prone.

Also included small styles changes from rubocop (spacing, whitespaces)