Create a placeholder source form for store credit
https://github.com/solidusio/solidus/issues/193 reports that if you
enable the Store Credit payment method in the admin, you will be unable
to access the new payments page for an order.

This PR creates a form so a ActionView::MissingTemplate exception isn't
raised. Additionally, it explains that create a store credit payment via
the admin is not currently supported.