Backport StoreCredit#amount_remaining Solidus PR
This commit backports work from an open [Solidus PR][1]

[1]: solidusio/solidus#1512

This migration adds the `amount_remaining` column to the
`spree_store_credit_events` table, and iterates over existing Store
Credit Events to calculate historical data for this column. This value
is also displayed in the admin, instead of the User's total remaining
store credit.