Replace :symbol with instance var in form_for
Current form_for is not compatible with `hints` because the helper does
not have access to the object. This commit replaces the :promotion
symbol with the @promotion instance variable.

Only difference between the old form_tag and the new form_tag is that
the new form_tag will have a `new_promotion` css class/html ID