Add Comparable to Spree::Money
This commit includes Comparable and implements `<=>`. Including
Comparable also gives us some basic comparison methods like `<`, `>`,
`.between?`, and `sort`.

We need to add an additional currency comparison because `::Money` will
automatically try to convert the second currency in order to be able to
compare it with the first. We don't want this to happen so we're going
to explicitly raise an error.

Conflicts:
	CHANGELOG.md
	core/lib/spree/money.rb
	core/spec/lib/spree/money_spec.rb